ui: enable remove datadoc from sidebar
